pyproject.toml
[tool.poetry]
name = "lois-lab-4"
version = "0.1.0"
description = "lab work 4 for LOIS"
authors = ["Ehor Hleoza <gleoza04@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.1.3"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
pyparsing = "^3.2.0"
python-intervals = "^1.10.0.post1"


[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
